- updates to playing queue gateway and repo - removed some interactors - updates to playing queue fragment - fixed shuffle not working as intended

+++ presentation/src/main/java/dev/olog/presentation/player/PlayerFragmentAdapter.kt
@@ -65,7 +65,7 @@ internal class PlayerFragmentAdapter(
         when (viewType) {
             R.layout.item_mini_queue -> {
                 viewHolder.setOnClickListener(this) { item, _, _ ->
-                    mediaProvider.skipToQueueItem(item.trackNumber.toLong())
+                    mediaProvider.skipToQueueItem(item.trackNumber.toInt())
                 }
                 viewHolder.setOnLongClickListener(this) { item, _, _ ->
                     navigator.toDialog(item, viewHolder.itemView)
